Files were moved to the web package
