+++ src/test/java/com/solvd/ecommerce/EcommerceTest.java
@@ -1,9 +1,9 @@
 package com.solvd.ecommerce;
 
 import com.qaprosoft.carina.core.foundation.IAbstractTest;
-import com.solvd.ecommerce.component.*;
-import com.solvd.ecommerce.page.*;
-import com.solvd.ecommerce.utils.CapabilitiesFactory;
+import com.solvd.ecommerce.web.component.*;
+import com.solvd.ecommerce.web.page.*;
+import com.solvd.ecommerce.web.utils.CapabilitiesFactory;
 import com.zebrunner.carina.utils.R;
 import org.testng.Assert;
 import org.testng.annotations.*;
